added OgImageService,logic moved from Worker to Service
