+++ app/workers/og_image_worker.rb
@@ -3,35 +3,6 @@ class OgImageWorker
   sidekiq_options retry: false
 
   def perform(project_id, page_id, page_title)
-    img = Magick::ImageList.new("#{Rails.root}/public/images/qkspace.png")
-
-    title = Magick::Draw.new
-    title.font = "#{Rails.root}/public/assets/Lato-Regular.woff"
-    title.pointsize = 46
-    title.fill = 'white'
-    title.gravity = Magick::CenterGravity
-    title.annotate(img, 0, 0, 0, 140, wrap_text(page_title))
-
-    dir = og_image_path(project_id, page_id)
-
-    FileUtils.mkdir_p(dir)
-
-    img.write("#{dir}/og-image.jpg")
-  end
-
-  def og_image_path(project_id, page_id)
-    if Rails.env.test?
-      "#{Rails.root}/public/images/opengraph/test/#{project_id}/#{page_id}"
-    else
-      "#{Rails.root}/public/images/opengraph/#{project_id}/#{page_id}"
-    end
-  end
-
-  private
-
-  def wrap_text(text, col = 50)
-    # Regular expression to limit the number of chars per line
-    # default 50 chars each line
-    text.gsub(/(.{1,#{col}})( +|$\n?)|(.{1,#{col}})/, "\\1\\3\n")
+    OgImageService.new(project_id, page_id, page_title).generate
   end
 end
