+++ config/application.rb
@@ -29,5 +29,6 @@ module Dostoevsky
 
     config.middleware.use DomainSpaceMiddleware
     config.active_job.queue_adapter = :sidekiq
+    Redis.exists_returns_integer =  true
   end
 end
